@@ -16,43 +16,34 @@
 // limitations under the License.
 
 using System;
-using System.Collections.Generic;
 using System.ComponentModel.Design;
 using System.Diagnostics.Contracts;
-using System.Globalization;
-using System.Net.Http;
 using System.IO;
-using System.Net;
 using System.Linq;
 using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
-using System.Windows.Forms;
-
-using Microsoft.VisualStudio.Debugger.Interop;
-using Microsoft.VisualStudio.Shell;
-using Microsoft.VisualStudio.Shell.Interop;
-
 using EnvDTE;
 using EnvDTE80;
-
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
 using Neon.Common;
 using Neon.IO;
 using Neon.SSH;
-using Neon.Windows;
-
 using Newtonsoft.Json.Linq;
-
+using RaspberryDebugger.Extensions;
+using RaspberryDebugger.Models.Connection;
+using RaspberryDebugger.Models.Project;
+using RaspberryDebugger.Models.VisualStudio;
 using Task = System.Threading.Tasks.Task;
 
-namespace RaspberryDebugger
+namespace RaspberryDebugger.Commands
 {
     /// <summary>
     /// Handles the <b>Start Debugging</b> command for Raspberry enabled projects.
     /// </summary>
     internal sealed class DebugStartCommand
     {
-        private DTE2    dte;
+        private readonly DTE2 dte;
 
         /// <summary>
         /// Command ID.
@@ -64,15 +55,8 @@ namespace RaspberryDebugger
         /// </summary>
         public static readonly Guid CommandSet = RaspberryDebuggerPackage.CommandSet;
 
-        /// <summary>
-        /// VS Package that provides this command, not null.
-        /// </summary>
-        private readonly AsyncPackage package;
-
-        /// <summary>
         /// Initializes a new instance of the <see cref="DebugStartCommand"/> class.
         /// Adds our command handlers for menu (commands must exist in the command table file)
-        /// </summary>
         /// <param name="package">Owner package, not null.</param>
         /// <param name="commandService">Command service to add command to, not null.</param>
         private DebugStartCommand(AsyncPackage package, OleMenuCommandService commandService)
@@ -82,24 +66,21 @@ namespace RaspberryDebugger
 
             ThreadHelper.ThrowIfNotOnUIThread();
 
-            this.package = package;
             this.dte     = (DTE2)Package.GetGlobalService(typeof(SDTE));
 
-            var menuCommandID = new CommandID(CommandSet, CommandId);
-            var menuItem      = new MenuCommand(this.Execute, menuCommandID);
+            var menuCommandId = new CommandID(CommandSet, CommandId);
+            var menuItem      = new MenuCommand(Execute, menuCommandId);
              
-            commandService.AddCommand(menuItem);
+            commandService?.AddCommand(menuItem);
         }
 
         /// <summary>
         /// Returns the command instance.
         /// </summary>
-        public static DebugStartCommand Instance { get; private set; }
-
-        /// <summary>
-        /// Gets the service provider from the owner package.
-        /// </summary>
-        private Microsoft.VisualStudio.Shell.IAsyncServiceProvider ServiceProvider => this.package;
+#pragma warning disable IDE0052 // Remove unread private members
+        // ReSharper disable once UnusedAutoPropertyAccessor.Local
+        private static DebugStartCommand Instance { get; set; }
+#pragma warning restore IDE0052 // Remove unread private members
 
         /// <summary>
         /// Initializes the singleton instance of the command.
@@ -121,9 +102,9 @@ namespace RaspberryDebugger
         /// </summary>
         /// <param name="sender">Event sender.</param>
         /// <param name="e">Event args.</param>
-#pragma warning disable VSTHRD100
+#pragma warning disable VSTHRD100 // Avoid async void methods
         private async void Execute(object sender, EventArgs e)
-#pragma warning restore VSTHRD100 
+#pragma warning restore VSTHRD100 // Avoid async void methods
         {
             await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
 
@@ -143,7 +124,7 @@ namespace RaspberryDebugger
 
             //-----------------------------------------------------------------
 
-            if (!await DebugHelper.PublishProjectWithUIAsync(dte, dte.Solution, project, projectProperties))
+            if (!await DebugHelper.PublishProjectWithUiAsync(dte, dte.Solution, project, projectProperties))
             {
                 return;
             }
@@ -155,20 +136,12 @@ namespace RaspberryDebugger
                 return;
             }
 
-            // Identify the most recent SDK installed on the workstation that has the same 
-            // major and minor version numbers as the project.  We'll ensure that the same
-            // SDK is installed on the Raspberry (further below).
-
-            var targetSdk = DebugHelper.GetTargetSdk(projectProperties);
-
-            if (targetSdk == null)
-            {
-                return;
-            }
-
+            var projectSettings = PackageHelper.GetProjectSettings(dte.Solution, project);
             // Establish a Raspberry connection to handle some things before we start the debugger.
-
-            var connection = await DebugHelper.InitializeConnectionAsync(connectionInfo, targetSdk, projectProperties, PackageHelper.GetProjectSettings(dte.Solution, project));
+            var connection = await DebugHelper.InitializeConnectionAsync(
+                connectionInfo, 
+                projectProperties, 
+                projectSettings);
 
             if (connection == null)
             {
@@ -178,71 +151,124 @@ namespace RaspberryDebugger
             using (connection)
             {
                 // Generate a temporary [launch.json] file and launch the debugger.
-
-                using (var tempFile = await CreateLaunchSettingsAsync(connectionInfo, projectProperties))
+                using (var tempFile = await CreateLaunchSettingsAsync(connectionInfo, projectProperties, projectSettings))
                 {
-                    dte.ExecuteCommand("DebugAdapterHost.Launch", $"/LaunchJson:\"{tempFile.Path}\"");
+                    try
+                    {
+                        dte.ExecuteCommand("DebugAdapterHost.Launch", $"/LaunchJson:\"{tempFile.Path}\"");
+                    }
+                    catch (Exception exception)
+                    {
+                        Console.WriteLine(exception);
+                    }
                 }
 
                 // Launch the browser for ASPNET apps if requested.  Note that we're going to do this
                 // on a background task to poll the Raspberry, waiting for the app to create the create
                 // the LISTENING socket.
+                if (!projectProperties.IsAspNet || !projectProperties.AspLaunchBrowser)
+                    return;
+                
+                var launchReady = false;
+                var foundWebServer = WebServer.None;
 
-                if (projectProperties.IsAspNet && projectProperties.AspLaunchBrowser)
-                {
-                    var baseUri     = $"http://{connectionInfo.Host}:{projectProperties.AspPort}";
-                    var launchReady = false;
+                await NeonHelper.WaitForAsync(async () =>
+                    {
+                        // The developer must have stopped debugging before the 
+                        // ASPNET application was able to begin servicing requests.
+                        if (dte.Mode != vsIDEMode.vsIDEModeDebug) return true;
 
-                    await NeonHelper.WaitForAsync(
-                        async () =>
+                        using (new CursorWait())
                         {
-                            if (dte.Mode != vsIDEMode.vsIDEModeDebug)
-                            {
-                                // The developer must have stopped debugging before the ASPNET
-                                // application was able to begin servicing requests.
-
-                                return true;
-                            }
-
                             try
                             {
-                                var appListeningScript =
-$@"
-if lsof -i -P -n | grep --quiet 'TCP \*:{projectProperties.AspPort} (LISTEN)' ; then
-    exit 0
-else
-    exit 1
-fi
-";
-                                var response = connection.SudoCommand(CommandBundle.FromScript(appListeningScript));
+                                var (found, webServer) =
+                                    await SearchForRunningWebServerAsync(projectProperties, projectSettings, connection);
 
-                                if (response.ExitCode != 0)
-                                {
-                                    return false;
-                                }
+                                // web server not found
+                                if (!found) return false;
 
-                            // Wait just a bit longer to give the application a chance to
-                            // perform any additional initialization.
-
-                            await Task.Delay(TimeSpan.FromSeconds(1));
-
+                                // take the found web server
+                                foundWebServer = webServer;
                                 launchReady = true;
+
                                 return true;
                             }
                             catch
                             {
                                 return false;
                             }
-                        },
-                        timeout: TimeSpan.FromSeconds(30),
-                        pollInterval: TimeSpan.FromSeconds(0.5));
+                        }
+                    },
+                    timeout: TimeSpan.FromSeconds(60),
+                    pollInterval: TimeSpan.FromSeconds(0.5));
 
-                    if (launchReady)
-                    {
-                        NeonHelper.OpenBrowser($"{baseUri}{projectProperties.AspRelativeBrowserUri}");
-                    }
-                }
+                if (!launchReady) return;
+
+                OpenWebBrowser(projectProperties, foundWebServer, connection);
             }
+        }
+
+        /// <summary>
+        /// Open web browser for debugging
+        /// </summary>
+        /// <param name="projectProperties">Related project properties</param>
+        /// <param name="foundWebServer">Active WebServer: Kestrel or Other (NGiNX, Apache, etc.)</param>
+        /// <param name="connection">LinuxSshProxy connection</param>
+        private static void OpenWebBrowser(
+            ProjectProperties projectProperties, 
+            WebServer foundWebServer, 
+            LinuxSshProxy connection)
+        {
+            // only '/' present or full relative uri
+            const int fullRelativeUri = 2;
+            var baseUri = $"http://{connection.Name}.local";
+
+            var relativeBrowserUri = projectProperties.AspRelativeBrowserUri.FirstOrDefault() == '/'
+                ? projectProperties.AspRelativeBrowserUri
+                : $"/{projectProperties.AspRelativeBrowserUri}";
+
+            var port = projectProperties.AspPort;
+
+            switch (foundWebServer)
+            {
+                // Apache, NGiNX, Kestrel and more... 
+                case WebServer.Other:
+                case WebServer.Kestrel:
+                    NeonHelper.OpenBrowser(relativeBrowserUri.Length < fullRelativeUri 
+                        ? foundWebServer == WebServer.Kestrel 
+                            ? $"{baseUri}:{port}" 
+                            : $"{baseUri}"
+                        : foundWebServer == WebServer.Kestrel 
+                            ? $"{baseUri}:{port}{relativeBrowserUri}" 
+                            : $"{baseUri}{relativeBrowserUri}");
+                    break;
+
+                // no running web server found
+                case WebServer.None:
+                default:
+                    break;
+            }
+        }
+
+        /// <summary>
+        /// Search for running web server
+        /// </summary>
+        /// <param name="projectProperties">ProjectProperties</param>
+        /// <param name="projectSettings"></param>
+        /// <param name="connection">LinuxSshProxy</param>
+        /// <returns>true if running</returns>
+        private static async Task<(bool, WebServer)> SearchForRunningWebServerAsync(ProjectProperties projectProperties,
+            ProjectSettings projectSettings,
+            LinuxSshProxy connection)
+        {
+            // Wait just a bit longer to give the application a 
+            // chance to perform any additional initialization.
+            await Task.Delay(TimeSpan.FromMilliseconds(125));
+
+            return projectSettings.UseWebServerProxy
+                ? ProxyWebServer.ListenFor(projectProperties.AspPort, connection, WebServer.Other)
+                : ProxyWebServer.ListenFor(projectProperties.AspPort, connection, WebServer.Kestrel);
         }
 
         /// <summary>
@@ -251,14 +277,16 @@ fi
         /// </summary>
         /// <param name="connectionInfo">The connection information.</param>
         /// <param name="projectProperties">The project properties.</param>
+        /// <param name="projectSettings"></param>
         /// <returns>The <see cref="TempFile"/> referencing the created launch file.</returns>
-        private async Task<TempFile> CreateLaunchSettingsAsync(ConnectionInfo connectionInfo, ProjectProperties projectProperties)
+        private async Task<TempFile> CreateLaunchSettingsAsync(ConnectionInfo connectionInfo,
+            ProjectProperties projectProperties, ProjectSettings projectSettings)
         {
             Covenant.Requires<ArgumentNullException>(connectionInfo != null, nameof(connectionInfo));
             Covenant.Requires<ArgumentNullException>(projectProperties != null, nameof(projectProperties));
 
             var systemRoot  = Environment.GetFolderPath(Environment.SpecialFolder.Windows);
-            var debugFolder = LinuxPath.Combine(PackageHelper.RemoteDebugBinaryRoot(connectionInfo.User), projectProperties.Name);
+            var debugFolder = LinuxPath.Combine(PackageHelper.RemoteDebugBinaryRoot(connectionInfo?.User), projectProperties?.Name);
 
             // NOTE:
             //
@@ -274,31 +302,43 @@ fi
             //
             // This means that we need add [program.dll] as the first argument, followed 
             // by the program arguments.
-
-            var args = new JArray();
-
-            args.Add(LinuxPath.Combine(debugFolder, projectProperties.AssemblyName + ".dll"));
-
-            foreach (var arg in projectProperties.CommandLineArgs)
+            var args = new JArray
             {
-                args.Add(arg);
+                LinuxPath.Combine(debugFolder, projectProperties?.AssemblyName + ".dll")
+            };
+
+            if (projectProperties?.CommandLineArgs != null)
+            {
+                foreach (var arg in projectProperties.CommandLineArgs)
+                {
+                    args.Add(arg);
+                }
             }
 
             var environmentVariables = new JObject();
 
-            foreach (var variable in projectProperties.EnvironmentVariables)
+            if (projectProperties?.EnvironmentVariables != null)
             {
-                environmentVariables.Add(variable.Key, variable.Value);
+                foreach (var variable in projectProperties.EnvironmentVariables)
+                {
+                    environmentVariables.Add(variable.Key, variable.Value);
+                }
             }
 
             // For ASPNET apps, set the [ASPNETCORE_URLS] environment variable
             // to [http://0.0.0.0:PORT] so that the app running on the Raspberry will
             // be reachable from the development workstation.  Note that we don't
             // support HTTPS at this time.
-
-            if (projectProperties.IsAspNet)
+            if (projectProperties?.IsAspNet == true)
             {
-                environmentVariables["ASPNETCORE_URLS"] = $"http://0.0.0.0:{projectProperties.AspPort}";
+                if (projectSettings.UseWebServerProxy)
+                {
+                    environmentVariables["ASPNETCORE_URLS"] = $"http://127.0.0.1:{projectProperties.AspPort}";
+                }
+                else
+                {
+                    environmentVariables["ASPNETCORE_URLS"] = $"http://0.0.0.0:{projectProperties.AspPort}";
+                }
             }
 
             // Construct the debug launch JSON file.
@@ -315,8 +355,8 @@ fi
                 new JObject
                 (
                     new JProperty("version", "0.2.1"),
-                    new JProperty("adapter", Path.Combine(systemRoot, "Sysnative", "OpenSSH", "ssh.exe")),
-                    new JProperty("adapterArgs", $"-i \"{connectionInfo.PrivateKeyPath}\" -o \"StrictHostKeyChecking no\" {connectionInfo.User}@{connectionInfo.Host} {PackageHelper.RemoteDebuggerPath} --interpreter=vscode {engineLogging}"),
+                    new JProperty("adapter", Path.Combine(systemRoot, "System32", "OpenSSH", "ssh.exe")),
+                    new JProperty("adapterArgs", $"-i \"{connectionInfo?.PrivateKeyPath}\" -o \"StrictHostKeyChecking no\" {connectionInfo?.User}@{connectionInfo?.Host} {PackageHelper.RemoteDebuggerPath} --interpreter=vscode {engineLogging}"),
                     new JProperty("configurations",
                         new JArray
                         (
@@ -338,10 +378,8 @@ fi
 
             var tempFile = new TempFile(".launch.json");
 
-            using (var stream = new FileStream(tempFile.Path, FileMode.CreateNew, FileAccess.ReadWrite))
-            {
-                await stream.WriteAsync(Encoding.UTF8.GetBytes(settings.ToString()));
-            }
+            using var stream = new FileStream(tempFile.Path, FileMode.CreateNew, FileAccess.ReadWrite);
+            await stream.WriteAsync(Encoding.UTF8.GetBytes(settings.ToString()));
 
             return tempFile;
         }
