@@ -14,23 +14,27 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics.Contracts;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Runtime.Serialization;
 using System.Threading.Tasks;
-
-using Neon.Common;
 using Neon.IO;
 using Neon.Net;
 using Neon.SSH;
+using RaspberryDebugger.Extensions;
+using RaspberryDebugger.Models.Connection;
+using RaspberryDebugger.Models.Project;
+using RaspberryDebugger.Models.Sdk;
 using Renci.SshNet.Common;
+using OperatingSystem = RaspberryDebugger.Models.Sdk.OperatingSystem;
 
-namespace RaspberryDebugger
+// ReSharper disable StringLiteralTypo
+
+namespace RaspberryDebugger.Connection
 {
     /// <summary>
     /// Implements a SSH connection to the remote Raspberry Pi.
@@ -59,25 +63,26 @@ namespace RaspberryDebugger
 
             try
             {
-                if (!NetHelper.TryParseIPv4Address(connectionInfo.Host, out var address))
+                if (!NetHelper.TryParseIPv4Address(connectionInfo?.Host, out var address))
                 {
-                    Log($"DNS lookup for: {connectionInfo.Host}");
+                    Log($"DNS lookup for: {connectionInfo?.Host}");
 
-                    address = (await Dns.GetHostAddressesAsync(connectionInfo.Host)).FirstOrDefault();
+                    var hosts = (await Dns.GetHostAddressesAsync(connectionInfo?.Host));
+                    address = hosts?.FirstOrDefault();
                 }
 
                 if (address == null)
                 {
-                    throw new ConnectionException(connectionInfo.Host, "DNS lookup failed.");
+                    throw new ConnectionException(connectionInfo?.Host, "DNS lookup failed.");
                 }
 
                 SshCredentials credentials;
 
-                if (string.IsNullOrEmpty(connectionInfo.PrivateKeyPath) || usePassword)
+                if (string.IsNullOrEmpty(connectionInfo?.PrivateKeyPath) || usePassword)
                 {
-                    Log($"[{connectionInfo.Host}]: Auth via username/password");
+                    Log($"[{connectionInfo?.Host}]: Auth via username/password");
 
-                    credentials = SshCredentials.FromUserPassword(connectionInfo.User, connectionInfo.Password);
+                    credentials = SshCredentials.FromUserPassword(connectionInfo?.User, connectionInfo?.Password);
                 }
                 else
                 {
@@ -86,7 +91,7 @@ namespace RaspberryDebugger
                     credentials = SshCredentials.FromPrivateKey(connectionInfo.User, File.ReadAllText(connectionInfo.PrivateKeyPath));
                 }
 
-                var connection = new Connection(connectionInfo.Host, address, connectionInfo, credentials, projectSettings);
+                var connection = new Connection(connectionInfo?.Host, address, connectionInfo, credentials, projectSettings);
 
                 connection.Connect(TimeSpan.Zero);
                 await connection.InitializeAsync();
@@ -99,21 +104,21 @@ namespace RaspberryDebugger
                     e.InnerException == null ||
                     e.InnerException.GetType() != typeof(SshAuthenticationException))
                 {
-                    RaspberryDebugger.Log.Exception(e, $"[{connectionInfo.Host}]");
+                    RaspberryDebugger.Log.Exception(e, $"[{connectionInfo?.Host}]");
                     throw;
                 }
 
-                if (string.IsNullOrEmpty(connectionInfo.PrivateKeyPath) ||
+                if (string.IsNullOrEmpty(connectionInfo?.PrivateKeyPath) ||
                     string.IsNullOrEmpty(connectionInfo.Password))
                 {
-                    RaspberryDebugger.Log.Exception(e, $"[{connectionInfo.Host}]: The connection must have a password or SSH private key.");
+                    RaspberryDebugger.Log.Exception(e, $"[{connectionInfo?.Host}]: The connection must have a password or SSH private key.");
                     throw;
                 }
 
                 RaspberryDebugger.Log.Warning($"[{connectionInfo.Host}]: SSH auth failed: Try using the password and reauthorizing the public key");
 
                 // SSH private key authentication didn't work.  This commonly happens
-                // after the user has reimaged the Raspberry.  It's likely that the
+                // after the user has re-imaged the Raspberry.  It's likely that the
                 // user has setup the same username/password, so we'll try logging in
                 // with those and just configure the current public key to be accepted
                 // on the Raspberry.
@@ -130,32 +135,34 @@ namespace RaspberryDebugger
                     var homeFolder = LinuxPath.Combine("/", "home", connectionInfo.User);
                     var publicKey  = File.ReadAllText(connectionInfo.PublicKeyPath).Trim();
                     var keyScript  =
-$@"
-mkdir -p {homeFolder}/.ssh
-touch {homeFolder}/.ssh/authorized_keys
+                        $@"
+                        mkdir -p {homeFolder}/.ssh
+                        touch {homeFolder}/.ssh/authorized_keys
 
-if ! grep --quiet '{publicKey}' {homeFolder}/.ssh/authorized_keys ; then
-    echo '{publicKey}' >> {homeFolder}/.ssh/authorized_keys
-    exit $?
-fi
+                        if ! grep --quiet '{publicKey}' {homeFolder}/.ssh/authorized_keys ; then
+                            echo '{publicKey}' >> {homeFolder}/.ssh/authorized_keys
+                            exit $?
+                        fi
 
-exit 0
-";
+                        exit 0
+                        ";
+
                     connection.ThrowOnError(connection.RunCommand(CommandBundle.FromScript(keyScript)));
                     return connection;
                 }
                 catch (Exception e2)
                 {
                     // We've done all we can.
-
                     RaspberryDebugger.Log.Exception(e2, $"[{connectionInfo.Host}]");
                     throw;
                 }
             }
             catch (Exception e)
             {
-                RaspberryDebugger.Log.Exception(e, $"[{connectionInfo.Host}]");
-                throw;
+                RaspberryDebugger.Log.Exception(e, $"[{connectionInfo?.Host}]");
+                Log($"[{connectionInfo?.Host}]: Cannot reach device.");
+
+                return null;
             }
         }
 
@@ -178,8 +185,8 @@ exit 0
         //---------------------------------------------------------------------
         // Instance members
 
-        private ConnectionInfo      connectionInfo;
-        private ProjectSettings     projectSettings;
+        private readonly ConnectionInfo  connectionInfo;
+        private readonly ProjectSettings projectSettings;
 
         /// <summary>
         /// Constructs a connection using a password.
@@ -256,9 +263,9 @@ exit 0
         {
             Covenant.Requires<ArgumentNullException>(commandResponse != null, nameof(commandResponse));
 
-            if (commandResponse.ExitCode != 0)
+            if (commandResponse?.ExitCode != 0)
             {
-                throw new ConnectionException(this, commandResponse.ErrorText);
+                throw new ConnectionException(this, commandResponse?.ErrorText);
             }
 
             return commandResponse;
@@ -277,7 +284,7 @@ exit 0
                 $"Connecting to [{Name}]...",
                 async () =>
                 {
-                    // Disabling this because it looks like SUDO passwork prompting is disabled
+                    // Disabling this because it looks like SUDO password prompting is disabled
                     // by default for Raspberry Pi OS.
 #if DISABLED
                     // This call ensures that SUDO password prompting is disabled and the
@@ -287,7 +294,6 @@ exit 0
 #endif
                     // We need to ensure that [unzip] is installed so that [LinuxSshProxy] command
                     // bundles will work.
-
                     Log($"[{Name}]: Checking for: [unzip]");
 
                     var response = SudoCommand("which unzip");
@@ -301,105 +307,95 @@ exit 0
                     }
 
                     // We're going to execute a script the gathers everything in a single operation for speed.
-
                     Log($"[{Name}]: Retrieving status");
 
                     var statusScript =
-$@"
-# This script will return the status information via STDOUT line-by-line
-# in this order:
-#
-# Chip Architecture
-# PATH environment variable
-# Unzip Installed (""unzip"" or ""unzip-missing"")
-# Debugger Installed (""debugger-installed"" or ""debugger-missing"")
-# List of installed SDKs names (e.g. 3.1.108) separated by commas
-# Raspberry Model like:     Raspberry Pi 4 Model B Rev 1.2
-# Raspberry Revision like:  c03112
-#
-# This script also ensures that the [/lib/dotnet] directory exists, that
-# it has reasonable permissions, and that the folder exists on the system
-# PATH and that DOTNET_ROOT points to the folder.
+                        $@"
+                        # This script will return the status information via STDOUT line-by-line
+                        # in this order:
+                        #
+                        # Chip Architecture
+                        # PATH environment variable
+                        # Unzip Installed (""unzip"" or ""unzip-missing"")
+                        # Debugger Installed (""debugger-installed"" or ""debugger-missing"")
+                        # List of installed SDKs names (e.g. 3.1.108) separated by commas
+                        # Raspberry Model like:     Raspberry Pi 4 Model B Rev 1.2
+                        # Raspberry Revision like:  c03112
+                        #
+                        # This script also ensures that the [/lib/dotnet] directory exists, that
+                        # it has reasonable permissions, and that the folder exists on the system
+                        # PATH and that DOTNET_ROOT points to the folder.
 
-# Set the SDK and debugger installation paths.
+                        # Set the SDK and debugger installation paths.
 
-DOTNET_ROOT={PackageHelper.RemoteDotnetFolder}
-DEBUGFOLDER={PackageHelper.RemoteDebuggerFolder}
+                        DOTNET_ROOT={PackageHelper.RemoteDotnetFolder}
+                        DEBUGFOLDER={PackageHelper.RemoteDebuggerFolder}
 
-# Get the chip architecture
+                        # Get the chip architecture
+                        uname -m
 
-uname -m
+                        # Get the current PATH
+                        echo $PATH
 
-# Get the current PATH
+                        # Detect whether [unzip] is installed.
+                        if which unzip &> /dev/nul ; then
+                            echo 'unzip'
+                        else
+                            echo 'unzip-missing'
+                        fi
 
-echo $PATH
+                        # Detect whether the [vsdbg] debugger is installed.
+                        if [ -d $DEBUGFOLDER ] ; then
+                            echo 'debugger-installed'
+                        else
+                            echo 'debugger-missing'
+                        fi
 
-# Detect whether [unzip] is installed.
+                        # List the SDK folders.  These folder names are the same as the
+                        # corresponding SDK name.  We'll list the files on one line
+                        # with the SDK names separated by commas.  We'll return a blank
+                        # line if the SDK directory doesn't exist.
+                        if [ -d $DOTNET_ROOT/sdk ] ; then
+                            ls -m $DOTNET_ROOT/sdk
+                        else
+                            echo ''
+                        fi
 
-if which unzip &> /dev/nul ; then
-    echo 'unzip'
-else
-    echo 'unzip-missing'
-fi
+                        # Output the Raspberry board model.
+                        cat /proc/cpuinfo | grep '^Model\s' | grep -o 'Raspberry.*$'
 
-# Detect whether the [vsdbg] debugger is installed.
+                        # Output the Raspberry board revision.
+                        cat /proc/cpuinfo | grep 'Revision\s' | grep -o '[0-9a-fA-F]*$'
 
-if [ -d $DEBUGFOLDER ] ; then
-    echo 'debugger-installed'
-else
-    echo 'debugger-missing'
-fi
+                        # Ensure that the [/lib/dotnet] folder exists, that it's on the
+                        # PATH and that DOTNET_ROOT are defined.
+                        mkdir -p /lib/dotnet
+                        chown root:root /lib/dotnet
+                        chmod 755 /lib/dotnet
 
-# List the SDK folders.  These folder names are the same as the
-# corresponding SDK name.  We'll list the files on one line
-# with the SDK names separated by commas.  We'll return a blank
-# line if the SDK directory doesn't exist.
+                        # Set these for the current session:
+                        export DOTNET_ROOT={PackageHelper.RemoteDotnetFolder}
+                        export PATH=$PATH:$DOTNET_ROOT
 
-if [ -d $DOTNET_ROOT/sdk ] ; then
-    ls -m $DOTNET_ROOT/sdk
-else
-    echo ''
-fi
+                        # and for future sessions too:
+                        if ! grep --quiet DOTNET_ROOT /etc/profile ; then
+                            echo """"                                >> /etc/profile
+                            echo ""#------------------------------"" >> /etc/profile
+                            echo ""# Raspberry Debugger:""           >> /etc/profile
+                            echo ""export DOTNET_ROOT=$DOTNET_ROOT"" >> /etc/profile
+                            echo ""export PATH=$PATH""               >> /etc/profile
+                            echo ""#------------------------------"" >> /etc/profile
+                        fi
+                        ";
 
-# Output the Raspberry board model.
-
-cat /proc/cpuinfo | grep '^Model\s' | grep -o 'Raspberry.*$'
-
-# Output the Raspberry board revision.
-
-cat /proc/cpuinfo | grep 'Revision\s' | grep -o '[0-9a-fA-F]*$'
-
-# Ensure that the [/lib/dotnet] folder exists, that it's on the
-# PATH and that DOTNET_ROOT are defined.
-
-mkdir -p /lib/dotnet
-chown root:root /lib/dotnet
-chmod 755 /lib/dotnet
-
-# Set these for the current session:
-
-export DOTNET_ROOT={PackageHelper.RemoteDotnetFolder}
-export PATH=$PATH:$DOTNET_ROOT
-
-# and for future sessions too:
-
-if ! grep --quiet DOTNET_ROOT /etc/profile ; then
-
-    echo """"                                >> /etc/profile
-    echo ""#------------------------------"" >> /etc/profile
-    echo ""# Raspberry Debugger:""           >> /etc/profile
-    echo ""export DOTNET_ROOT=$DOTNET_ROOT"" >> /etc/profile
-    echo ""export PATH=$PATH""               >> /etc/profile
-    echo ""#------------------------------"" >> /etc/profile
-fi
-";
                     Log($"[{Name}]: Fetching status");
 
                     response = ThrowOnError(SudoCommand(CommandBundle.FromScript(statusScript)));
 
+                    // ReSharper disable once ConvertToUsingDeclaration
                     using (var reader = new StringReader(response.OutputText))
                     {
-                        var architecture = await reader.ReadLineAsync();
+                        var processor    = await reader.ReadLineAsync();
                         var path         = await reader.ReadLineAsync();
                         var hasUnzip     = await reader.ReadLineAsync() == "unzip";
                         var hasDebugger  = await reader.ReadLineAsync() == "debugger-installed";
@@ -409,54 +405,66 @@ fi
 
                         revision = revision.Trim();     // Remove any whitespace at the end.
 
-                        Log($"[{Name}]: architecture: {architecture}");
-                        Log($"[{Name}]: path:         {path}");
-                        Log($"[{Name}]: unzip:        {hasUnzip}");
-                        Log($"[{Name}]: debugger:     {hasDebugger}");
-                        Log($"[{Name}]: sdks:         {sdkLine}");
-                        Log($"[{Name}]: model:        {model}");
-                        Log($"[{Name}]: revision:     {revision}");
+                        Log($"[{Name}]: processor: {processor}");
+                        Log($"[{Name}]: path:      {path}");
+                        Log($"[{Name}]: unzip:     {hasUnzip}");
+                        Log($"[{Name}]: debugger:  {hasDebugger}");
+                        Log($"[{Name}]: sdks:      {sdkLine}");
+                        Log($"[{Name}]: model:     {model}");
+                        Log($"[{Name}]: revision:  {revision}");
+
+                        var osBitness = SdkArchitecture.Unknown;
+
+                        if (OperatingSystem.Bitness32.Any(bitness => processor.Contains(bitness)))
+                        {
+                            osBitness = SdkArchitecture.Arm32;
+                        }
+
+                        if (OperatingSystem.Bitness64.Any(bitness => processor.Contains(bitness)))
+                        {
+                            osBitness = SdkArchitecture.Arm64;
+                        }
 
                         // Convert the comma separated SDK names into a [PiSdk] list.
-
                         var sdks = new List<Sdk>();
 
-                        foreach (var sdkName in sdkLine.Split(new char[] { ',' }, StringSplitOptions.RemoveEmptyEntries).Select(sdk => sdk.Trim()))
+                        foreach (var sdkName in sdkLine
+                                     .Split(new[] { ',' }, StringSplitOptions.RemoveEmptyEntries)
+                                     .Select(sdk => sdk.Trim()))
                         {
-                            // $todo(jefflill): We're only supporting 32-bit SDKs at this time.
-
-                            var sdkCatalogItem = PackageHelper.SdkCatalog.Items.SingleOrDefault(item => item.Name == sdkName && item.Architecture == SdkArchitecture.ARM32);
+                            var sdkCatalogItem = PackageHelper.SdkCatalog.Items
+                                .SingleOrDefault(item => item.Link.Contains(sdkName) && item.Architecture == osBitness);
 
                             if (sdkCatalogItem != null)
                             {
-                                sdks.Add(new Sdk(sdkName, sdkCatalogItem.Version));
+                                sdks.Add(new Sdk(sdkName, osBitness));
                             }
                             else
                             {
-                                LogWarning($".NET SDK [{sdkName}] is present on [{Name}] but is not known to the RaspberryDebugger extension.  Consider updating the extension.");
+                                LogWarning($".NET SDK [{sdkName}] is present on [{Name}] but is not known to the RaspberryDebugger extension. Consider updating the extension.");
                             }
                         }
 
                         PiStatus = new Status(
-                            architecture:  architecture,
+                            processor:     processor,
                             path:          path,
                             hasUnzip:      hasUnzip,
                             hasDebugger:   hasDebugger,
                             installedSdks: sdks,
                             model:         model,
-                            revision:      revision
+                            revision:      revision,
+                            architecture:  osBitness
                         );
                     }
                 });
 
             // Create and configure an SSH key for this connection if one doesn't already exist.
-
             if (string.IsNullOrEmpty(connectionInfo.PrivateKeyPath) || !File.Exists(connectionInfo.PrivateKeyPath))
             {
                 await PackageHelper.ExecuteWithProgressAsync("Creating SSH keys...",
                     async () =>
                     {
-                        // Create a 2048-bit private key with no passphrase on the Raspberry
+                        // Create a 2048-bit private key with no pass-phrase on the Raspberry
                         // and then download it to our keys folder.  The key file name will
                         // be the host name of the Raspberry.
 
@@ -472,26 +480,24 @@ fi
                         try
                         {
                             var createKeyScript =
-$@"
-# Create the key pair
+                                $@"
+                                # Create the key pair
+                                if ! ssh-keygen -t rsa -b 2048 -P '' -C '{workstationUser}@{workstationName}' -f {tempPrivateKeyPath} -m pem ; then
+                                    exit 1
+                                fi
 
-if ! ssh-keygen -t rsa -b 2048 -P '' -C '{workstationUser}@{workstationName}' -f {tempPrivateKeyPath} -m pem ; then
-    exit 1
-fi
+                                # Append the public key to the user's [authorized_keys] file to enable it.
+                                mkdir -p {homeFolder}/.ssh
+                                touch {homeFolder}/.ssh/authorized_keys
+                                cat {tempPublicKeyPath} >> {homeFolder}/.ssh/authorized_keys
 
-# Append the public key to the user's [authorized_keys] file to enable it.
+                                exit 0
+                                ";
 
-mkdir -p {homeFolder}/.ssh
-touch {homeFolder}/.ssh/authorized_keys
-cat {tempPublicKeyPath} >> {homeFolder}/.ssh/authorized_keys
-
-exit 0
-";
                             ThrowOnError(RunCommand(CommandBundle.FromScript(createKeyScript)));
 
                             // Download the public and private keys, persist them to the workstation
                             // and then update the connection info.
-
                             var connections            = PackageHelper.ReadConnections();
                             var existingConnectionInfo = connections.SingleOrDefault(c => c.Name == connectionInfo.Name);
                             var publicKeyPath          = Path.Combine(PackageHelper.KeysFolder, $"{connectionInfo.Name}.pub");
@@ -514,12 +520,12 @@ exit 0
                         finally
                         {
                             // Delete the temporary key files on the Raspberry.
-
                             var removeKeyScript =
-$@"
-rm -f {tempPrivateKeyPath}
-rm -f {tempPublicKeyPath}
-";
+                                $@"
+                                rm -f {tempPrivateKeyPath}
+                                rm -f {tempPublicKeyPath}
+                                ";
+
                             ThrowOnError(SudoCommand(CommandBundle.FromScript(removeKeyScript)));
                         }
 
@@ -529,129 +535,81 @@ rm -f {tempPublicKeyPath}
         }
 
         /// <summary>
-        /// Installs the specified .NET Core SDK on the Raspberry if it's not already installed.
+        /// Download and install actual (latest) SDK
         /// </summary>
-        /// <param name="sdkVersion">The SDK version.</param>
-        /// <returns><c>true</c> on success.</returns>
-        public async Task<bool> InstallSdkAsync(string sdkVersion)
+        /// <returns>true if successful</returns>
+        public async Task<bool> SetupSdkAsync()
         {
-            Covenant.Requires<ArgumentNullException>(!string.IsNullOrEmpty(sdkVersion), nameof(sdkVersion));
+            if (IsSdkPresent()) return true;
 
-            // $todo(jefflill):
-            //
-            // Note that we're going to install that standalone SDK for the SDK
-            // version rather than the SDK that shipped with Visual Studio.  I'm
-            // assuming that the Visual Studio SDKs might have extra stuff we don't
-            // need and it's also possible that the Visual Studio SDK for the SDK
-            // version may not have shipped yet.
-            //
-            // We may want to re-evaluate this in the future.
+            var targetSdk = ReadActualSdkCatalogItem();
 
-            if (PiStatus.InstalledSdks.Any(sdk => sdk.Version == sdkVersion))
-            {
-                return await Task.FromResult(true);    // Already installed
-            }
+            return await DownloadSdkAsync(targetSdk) && 
+                   await InstallSdkAsync(targetSdk);
+        }
 
-            LogInfo($".NET Core SDK [v{sdkVersion}] is not installed.");
-
-            // Locate the standalone SDK for the request .NET version.
-
-            var targetSdk = PackageHelper.SdkGoodCatalog.Items.SingleOrDefault(item => item.IsStandalone && item.Version == sdkVersion && item.Architecture == SdkArchitecture.ARM32);
-
+        /// <summary>
+        /// Downloads the .NET Core SDK on the Raspberry.
+        /// The Raspberry architecture is driving the installation - the newest .NET Core version is taken
+        /// </summary>
+        /// <returns><c>true</c> on success.</returns>
+        private async Task<bool> DownloadSdkAsync(SdkCatalogItem targetSdk)
+        {
             if (targetSdk == null)
             {
-                // Fall back to the Visual Studio SDK, if there is one.
-
-                targetSdk = PackageHelper.SdkGoodCatalog.Items.SingleOrDefault(item => item.Version == sdkVersion);
-                LogInfo($"Cannot find standalone SDK for [{sdkVersion}] for falling back to [{targetSdk.Name}], version [v{targetSdk.Version}].");
-            }
-
-            if (targetSdk == null)
-            {
-                LogError($"RasberryDebug is unaware of .NET Core SDK [v{sdkVersion}].");
-                LogError($"Try updating the RasberryDebug extension or report this issue at:");
-                LogError($"https://github.com/nforgeio/RaspberryDebugger/issues");
+                LogError("RasberryDebug is unaware of .NET Core SDK.");
+                LogError("Try updating the RasberryDebug extension or report this issue at:");
+                LogError("https://github.com/nforgeio/RaspberryDebugger/issues");
 
                 return await Task.FromResult(false);
             }
+            else
+            {
+                LogInfo($".NET Core SDK [v{targetSdk.Release}] is not installed.");
+            }
 
             // Install the SDK.
+            LogInfo($"Downlaoding SDK v{targetSdk.Release}");
+            
+            var downloadSdkInfo = 
+                $"Download SDK for .NET v{targetSdk.Release} " +
+                $"({targetSdk.Architecture.GetAttributeOfType<EnumMemberAttribute>().Value}) on Raspberry...";
 
-            LogInfo($"Installing SDK v{targetSdk.Version}");
-
-            return await PackageHelper.ExecuteWithProgressAsync<bool>($"Download and install SDK for .NET v{targetSdk.Version} on Raspberry...",
+            return await PackageHelper.ExecuteWithProgressAsync(downloadSdkInfo,
                 async () =>
                 {
-                    var installScript =
-$@"
-export DOTNET_ROOT={PackageHelper.RemoteDotnetFolder}
+                    var downloadScript =
+                        $@"
+                        # Ensure that the packages required by .NET Core are installed:
+                        # https://docs.microsoft.com/en-us/dotnet/core/install/linux-debian#dependencies
+                        if ! apt-get update ; then
+                            exit 1
+                        fi
 
-# Ensure that the packages required by .NET Core are installed:
-#
-#       https://docs.microsoft.com/en-us/dotnet/core/install/linux-debian#dependencies
+                        if ! apt-get install -yq libc6 libgcc1 libgssapi-krb5-2 libicu-dev libssl1.1 libstdc++6 zlib1g libgdiplus ; then
+                            exit 1
+                        fi
 
-if ! apt-get update ; then
-    exit 1
-fi
+                        # Remove any existing SDK download.  This might be 
+                        # present if a previous installation attempt failed.
+                        if ! rm -f /tmp/dotnet-sdk.tar.gz ; then
+                            exit 1
+                        fi
 
-if ! apt-get install -yq libc6 libgcc1 libgssapi-krb5-2 libicu-dev libssl1.1 libstdc++6 zlib1g libgdiplus ; then
-    exit 1
-fi
+                        # Download the SDK installation file to a temporary file.
+                        if ! wget --quiet -O /tmp/dotnet-sdk.tar.gz {targetSdk.Link} ; then
+                            exit 1
+                        fi
 
-# Remove any existing SDK download.  This might be present if a
-# previous installation attempt failed.
+                        exit 0
+                        ";
 
-if ! rm -f /tmp/dotnet-sdk.tar.gz ; then
-    exit 1
-fi
-
-# Download the SDK installation file to a temporary file.
-
-if ! wget --quiet -O /tmp/dotnet-sdk.tar.gz {targetSdk.Link} ; then
-    exit 1
-fi
-
-# Verify the SHA512.
-
-orgDir=$cwd
-cd /tmp
-
-if ! echo '{targetSdk.SHA512}  dotnet-sdk.tar.gz' | sha512sum --check - ; then
-    cd $orgDir
-    exit 1
-fi
-
-cd $orgDir
-
-# Make sure the installation directory exists.
-
-if ! mkdir -p $DOTNET_ROOT ; then
-    exit 1
-fi
-
-# Unpack the SDK to the installation directory.
-
-if ! tar -zxf /tmp/dotnet-sdk.tar.gz -C $DOTNET_ROOT --no-same-owner ; then
-    exit 1
-fi
-
-# Remove the temporary installation file.
-
-if ! rm /tmp/dotnet-sdk.tar.gz ; then
-    exit 1
-fi
-
-exit 0
-";
                     try
                     {
-                        var response = SudoCommand(CommandBundle.FromScript(installScript));
+                        var response = SudoCommand(CommandBundle.FromScript(downloadScript));
 
                         if (response.ExitCode == 0)
                         {
-                            // Add the newly installed SDK to the list of installed SDKs.
-
-                            PiStatus.InstalledSdks.Add(new Sdk(targetSdk.Name, targetSdk.Version));
                             return await Task.FromResult(true);
                         }
                         else
@@ -669,10 +627,117 @@ exit 0
         }
 
         /// <summary>
+        /// Installs the .NET Core SDK on the Raspberry if it's not already installed.
+        /// The Raspberry architecture is driving the installation - the newest .NET Core version is taken
+        /// </summary>
+        /// <returns><c>true</c> on success.</returns>
+        private async Task<bool> InstallSdkAsync(SdkCatalogItem targetSdk)
+        {
+            // Install the SDK.
+            LogInfo($"Installing SDK v{targetSdk.Release}");
+            
+            var installSdkInfo = 
+                $"Install SDK for .NET v{targetSdk.Release} " +
+                $"({targetSdk.Architecture.GetAttributeOfType<EnumMemberAttribute>().Value}) on Raspberry...";
+
+            return await PackageHelper.ExecuteWithProgressAsync(installSdkInfo,
+                async () =>
+                {
+                    var installScript =
+                        $@"
+                        export DOTNET_ROOT={PackageHelper.RemoteDotnetFolder}
+
+                        # Verify the SHA512.
+                        orgDir=$cwd
+                        cd /tmp
+
+                        if ! echo '{targetSdk.Sha512} dotnet-sdk.tar.gz' | sha512sum --check - ; then
+                            cd $orgDir
+                            exit 1
+                        fi
+
+                        cd $orgDir
+
+                        # Make sure the installation directory exists.
+                        if ! mkdir -p $DOTNET_ROOT ; then
+                            exit 1
+                        fi
+
+                        # Unpack the SDK to the installation directory.
+                        if ! tar -zxf /tmp/dotnet-sdk.tar.gz -C $DOTNET_ROOT --no-same-owner ; then
+                            exit 1
+                        fi
+
+                        # Remove the temporary installation file.
+                        if ! rm /tmp/dotnet-sdk.tar.gz ; then
+                            exit 1
+                        fi
+
+                        exit 0
+                        ";
+
+                    try
+                    {
+                        var response = SudoCommand(CommandBundle.FromScript(installScript));
+
+                        if (response.ExitCode == 0)
+                        {
+                            // Add the newly installed SDK to the list of installed SDKs.
+                            PiStatus.InstalledSdks.Add(new Sdk(targetSdk.Name, targetSdk.Architecture));
+                            return await Task.FromResult(true);
+                        }
+                        else
+                        {
+                            LogError(response.AllText);
+                            return await Task.FromResult(false);
+                        }
+                    }
+                    catch (Exception e)
+                    {
+                        LogException(e);
+                        return await Task.FromResult(false);
+                    }
+                });
+        }
+
+        /// <summary>
+        /// Is any .NET Core SDK installed on the Raspberry Pi.
+        /// The Raspberry architecture is driving the installation - the newest .NET Core version is taken
+        /// </summary>
+        /// <returns><c>true</c> on success.</returns>
+        private bool IsSdkPresent()
+        {
+            var sdkOnPi = PiStatus.InstalledSdks
+                .OrderByDescending(name => name.Name)
+                .FirstOrDefault();
+
+            var sdkOnPiVersion = sdkOnPi?.Name ?? string.Empty;
+            var sdkOnPiArchitecture = sdkOnPi?.Architecture ?? PiStatus.Architecture;
+
+            return PiStatus.InstalledSdks.Any(sdk => sdk.Name == sdkOnPiVersion && 
+                                                     sdk.Architecture == sdkOnPiArchitecture);
+        }
+
+        /// <summary>
+        /// Read actual (latest) SDK from catalog
+        /// </summary>
+        /// <returns>Latest target SDK</returns>
+        private SdkCatalogItem ReadActualSdkCatalogItem()
+        {
+            // Locate the standalone SDK for the request .NET version.
+            // Figure out the latest SDK version - Microsoft versioning: the highest number
+            var targetSdk = PackageHelper.SdkCatalog.Items
+                .OrderByDescending(item => item.Name)
+                .FirstOrDefault(item => item.Architecture == PiStatus.Architecture);
+
+            return targetSdk;
+        }
+
+        /// <summary>
         /// Installs the <b>vsdbg</b> debugger on the Raspberry if it's not already installed.
         /// </summary>
         /// <returns><c>true</c> on success.</returns>
-        public async Task<bool> InstallDebuggerAsync()
+        public async Task<bool> SetupDebuggerAsync()
         {
             if (PiStatus.HasDebugger)
             {
@@ -681,17 +746,18 @@ exit 0
 
             LogInfo($"Installing VSDBG to: [{PackageHelper.RemoteDebuggerFolder}]");
 
-            return await PackageHelper.ExecuteWithProgressAsync<bool>($"Installing [vsdbg] debugger...",
+            return await PackageHelper.ExecuteWithProgressAsync("Installing [vsdbg] debugger...",
                 async () =>
                 {
                     var installScript =
-$@"
-if ! curl -sSL https://aka.ms/getvsdbgsh | /bin/sh /dev/stdin -v latest -l {PackageHelper.RemoteDebuggerFolder} ; then
-    exit 1
-fi
+                        $@"
+                        if ! curl -sSL https://aka.ms/getvsdbgsh | /bin/sh /dev/stdin -v latest -l {PackageHelper.RemoteDebuggerFolder} ; then
+                            exit 1
+                        fi
 
-exit 0
-";
+                        exit 0
+                        ";
+
                     try
                     {
                         var response = SudoCommand(CommandBundle.FromScript(installScript));
@@ -722,15 +788,15 @@ exit 0
         /// any existing files.
         /// </summary>
         /// <param name="programName">The program name</param>
-        /// <param name="assemblyName">The addembly name.</param>
+        /// <param name="assemblyName">The assembly name.</param>
         /// <param name="publishedBinaryFolder">Path to the workstation folder holding the program files.</param>
         /// <returns><c>true</c> on success.</returns>
         public async Task<bool> UploadProgramAsync(string programName, string assemblyName, string publishedBinaryFolder)
         {
             Covenant.Requires<ArgumentNullException>(!string.IsNullOrEmpty(programName), nameof(programName));
-            Covenant.Requires<ArgumentException>(!programName.Contains(' '), nameof(programName));
+            Covenant.Requires<ArgumentException>(!(programName ?? string.Empty).Contains(' '), nameof(programName));
             Covenant.Requires<ArgumentNullException>(!string.IsNullOrEmpty(assemblyName), nameof(assemblyName));
-            Covenant.Requires<ArgumentException>(!assemblyName.Contains(' '), nameof(assemblyName));
+            Covenant.Requires<ArgumentException>(!(assemblyName ?? string.Empty).Contains(' '), nameof(assemblyName));
             Covenant.Requires<ArgumentNullException>(!string.IsNullOrEmpty(publishedBinaryFolder), nameof(publishedBinaryFolder));
             Covenant.Requires<ArgumentNullException>(Directory.Exists(publishedBinaryFolder), nameof(publishedBinaryFolder));
 
@@ -743,47 +809,41 @@ exit 0
             if (!string.IsNullOrEmpty(projectSettings.TargetGroup))
             {
                 groupScript =
-$@"
-# Add the program assembly to the user specified target group (if any).  This
-# defaults to [gpio] so users will be able to access the GPIO pins.
-
-if ! chgrp {projectSettings.TargetGroup} {debugFolder}/{assemblyName} ; then
-    exit 1
-fi
-";
+                    $@"
+                    # Add the program assembly to the user specified target group (if any).
+                    # This defaults to [gpio] so users will be able to access the GPIO pins.
+                    if ! chgrp {projectSettings.TargetGroup} {debugFolder}/{assemblyName} ; then
+                        exit 1
+                    fi
+                    ";
             }
 
             var uploadScript =
-$@"
+                $@"
+                # Ensure that the debug folder exists.
+                if ! mkdir -p {debugFolder} ; then
+                    exit 1
+                fi
 
-# Ensure that the debug folder exists.
+                # Clear all existing program files.
+                if ! rm -rf {debugFolder}/* ; then
+                    exit 1
+                fi
 
-if ! mkdir -p {debugFolder} ; then
-    exit 1
-fi
+                # Unzip the binary and other files to the debug folder.
+                if ! unzip program.zip -d {debugFolder} ; then
+                    exit 1
+                fi
 
-# Clear all existing program files.
+                # The program assembly needs execute permissions.
+                if ! chmod 770 {debugFolder}/{assemblyName} ; then
+                    exit 1
+                fi
+                {groupScript}
+                exit 0
+                ";
 
-if ! rm -rf {debugFolder}/* ; then
-    exit 1
-fi
-
-# Unzip the binary and other files to the debug folder.
-
-if ! unzip program.zip -d {debugFolder} ; then
-    exit 1
-fi
-
-# The program assembly needs execute permissions.
-
-if ! chmod 770 {debugFolder}/{assemblyName} ; then
-    exit 1
-fi
-{groupScript}
-exit 0
-";
             // I'm not going to do a progress dialog because this should be fast.
-
             try
             {
                 LogInfo($"Uploading program to: [{debugFolder}]");
@@ -796,7 +856,7 @@ exit 0
 
                 if (response.ExitCode == 0)
                 {
-                    LogInfo($"Program uploaded");
+                    LogInfo("Program uploaded");
                     return await Task.FromResult(true);
                 }
                 else
