@@ -15,20 +15,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-using Neon.Common;
-using BrightIdeasSoftware;
-
-namespace RaspberryDebugger
+namespace RaspberryDebugger.Dialogs
 {
     partial class ConnectionDialog
     {
