@@ -17,28 +17,22 @@
 
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
 using System.Linq;
 using System.Net;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-
-using Neon.Common;
 using Neon.Net;
+using RaspberryDebugger.Models.Connection;
 
-namespace RaspberryDebugger
+namespace RaspberryDebugger.Dialogs
 {
     /// <summary>
     /// Implements the Add/Remove connection dialogs.
     /// </summary>
     internal partial class ConnectionDialog : Form
     {
-        private const char passwordChar = '•';
+        private const char PasswordChar = '•';
 
-        private List<ConnectionInfo>    existingConnections;
+        private readonly List<ConnectionInfo> existingConnections;
 
         /// <summary>
         /// Constructor.
@@ -55,23 +49,28 @@ namespace RaspberryDebugger
             this.existingConnections = existingConnections;
 
             // Initialize the controls on load.
-
             this.Load += (s, a) =>
             {
                 hostTextBox.Text             = connectionInfo.Host;
                 portTextBox.Text             = connectionInfo.Port.ToString();
                 userTextBox.Text             = connectionInfo.User;
                 passwordTextBox.Text         = connectionInfo.Password;
-                passwordTextBox.PasswordChar = passwordChar;
+                passwordTextBox.PasswordChar = PasswordChar;
                 showPasswordCheckBox.Checked = false;
                 instructionsTextBox.Visible  = string.IsNullOrEmpty(connectionInfo.PrivateKeyPath);
             };
         }
 
+        public sealed override string Text
+        {
+            get => base.Text;
+            set => base.Text = value;
+        }
+
         /// <summary>
         /// Returns the connection being created or edited.
         /// </summary>
-        internal ConnectionInfo ConnectionInfo { get; private set; }
+        private ConnectionInfo ConnectionInfo { get; }
 
         /// <summary>
         /// Handles the OK button.
@@ -95,7 +94,7 @@ namespace RaspberryDebugger
                 return;
             }
 
-            if (!IPAddress.TryParse(hostText, out var address) && !NetHelper.IsValidHost(hostText))
+            if (!IPAddress.TryParse(hostText, out _) && !NetHelper.IsValidHost(hostText))
             {
                 portTextBox.Focus();
                 portTextBox.SelectAll();
@@ -183,9 +182,6 @@ namespace RaspberryDebugger
                 return;
             }
 
-            hasWhitespace = false;
-            hasQuote      = false;
-
             foreach (var ch in passwordText)
             {
                 if (char.IsWhiteSpace(ch))
@@ -244,9 +240,7 @@ namespace RaspberryDebugger
 
             try
             {
-                using (await Connection.ConnectAsync(testConnectionInfo))
-                {
-                }
+                using (await Connection.Connection.ConnectAsync(testConnectionInfo)) { }
             }
             catch
             {
@@ -290,7 +284,7 @@ namespace RaspberryDebugger
         /// <param name="args">The arguments.</param>
         private void showPasswordCheckBox_CheckedChanged(object sender, EventArgs args)
         {
-            passwordTextBox.PasswordChar = showPasswordCheckBox.Checked ? (char)0 : passwordChar;
+            passwordTextBox.PasswordChar = showPasswordCheckBox.Checked ? (char)0 : PasswordChar;
         }
     }
 }
