@@ -16,20 +16,9 @@
 // limitations under the License.
 
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
-using System.Linq;
-using System.Net;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-using Neon.Common;
-using Neon.Net;
-
-namespace RaspberryDebugger
+namespace RaspberryDebugger.Dialogs
 {
     /// <summary>
     /// Edits a connection's SSH key pair.
