@@ -14,32 +14,22 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
+using RaspberryDebugger.Models.Project;
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics.Contracts;
-using System.Drawing;
-using System.Data;
 using System.Linq;
-using System.Net;
-using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-using Neon.Common;
-using Neon.Net;
-
-namespace RaspberryDebugger
+namespace RaspberryDebugger.Dialogs
 {
     /// <summary>
-    /// Allows the user to edit the Rasparry related settings for a project.
+    /// Allows the user to edit the Raspberry related settings for a project.
     /// </summary>
     internal partial class SettingsDialog : Form
     {
-        private ProjectSettings             projectSettings;
-        private Dictionary<string, int>     connectionNameToIndex;
+        private readonly ProjectSettings projectSettings;
 
         /// <summary>
         /// Constructor.
@@ -48,23 +38,20 @@ namespace RaspberryDebugger
         public SettingsDialog(ProjectSettings projectSettings)
         {
             Covenant.Requires<ArgumentNullException>(projectSettings != null);
-
             this.projectSettings = projectSettings;
 
             InitializeComponent();
 
             // The instructions include "\r\n" sequences that need to be replaced with
             // actual CR/LF characters.
-
             instructionsTextBox.Text = Regex.Unescape(instructionsTextBox.Text);
 
             // Initialize the combo box with the available connections and select
             // the current one.
-
-            connectionNameToIndex = new Dictionary<string, int>(StringComparer.InvariantCultureIgnoreCase);
+            var connectionNameToIndex = new Dictionary<string, int>(StringComparer.InvariantCultureIgnoreCase);
 
             var connections = PackageHelper.ReadConnections(disableLogging: true);
-            var index       = 0;
+            var index = 0;
 
             targetComboBox.Items.Clear();
 
@@ -80,7 +67,7 @@ namespace RaspberryDebugger
                 connectionNameToIndex.Add(connection.Name, index++);
             }
 
-            if (!projectSettings.EnableRemoteDebugging)
+            if (projectSettings is { EnableRemoteDebugging: false })
             {
                 targetComboBox.SelectedIndex = connectionNameToIndex[ProjectSettings.DisabledConnectionName];
             }
@@ -88,27 +75,32 @@ namespace RaspberryDebugger
             {
                 // If the connection named in the settings exists select it,
                 // otherwise select the default.
+                var selectedConnection = connections.FirstOrDefault(connection => projectSettings != null
+                    && connection.Name.Equals(projectSettings.RemoteDebugTarget, StringComparison.InvariantCultureIgnoreCase));
 
-                var selectedConnection = connections.FirstOrDefault(connection => connection.Name.Equals(projectSettings.RemoteDebugTarget, StringComparison.InvariantCultureIgnoreCase));
-
-                if (projectSettings.RemoteDebugTarget == null || selectedConnection == null)
+                if (projectSettings != null && (projectSettings.RemoteDebugTarget == null || selectedConnection == null))
                 {
                     targetComboBox.SelectedIndex = connectionNameToIndex[ProjectSettings.DefaultConnectionName];
                 }
                 else
                 {
-                    targetComboBox.SelectedIndex = connectionNameToIndex[selectedConnection.Name];
+                    if (selectedConnection != null)
+                    {
+                        targetComboBox.SelectedIndex = connectionNameToIndex[selectedConnection.Name];
+                    }
                 }
             }
 
+            if (projectSettings == null) return;
             targetGroup.Text = projectSettings.TargetGroup;
+            webServerProxyCheck.Checked = projectSettings.UseWebServerProxy;
         }
 
         /// <summary>
         /// Handles OK button clicks.
         /// </summary>
         /// <param name="sender">The sender</param>
-        /// <param name="args"The arguments</param>
+        /// <param name="args">Event arguments</param>
         private void okButton_Click(object sender, EventArgs args)
         {
             var selectedItem = (string)(targetComboBox.SelectedItem);
@@ -123,13 +115,13 @@ namespace RaspberryDebugger
                 case ProjectSettings.DefaultConnectionName:
 
                     projectSettings.EnableRemoteDebugging = true;
-                    projectSettings.RemoteDebugTarget     = null;   // NULL means default
+                    projectSettings.RemoteDebugTarget = null;   // NULL means default
                     break;
 
                 default:
 
                     projectSettings.EnableRemoteDebugging = true;
-                    projectSettings.RemoteDebugTarget     = selectedItem;
+                    projectSettings.RemoteDebugTarget = selectedItem;
                     break;
             }
 
@@ -141,8 +133,8 @@ namespace RaspberryDebugger
                 targetGroup.SelectAll();
 
                 MessageBoxEx.Show(
-                    $"Invalid Linux group name: group names should not include spaces.",
-                    $"Target Group Error",
+                    "Invalid Linux group name: group names should not include spaces.",
+                    "Target Group Error",
                     MessageBoxButtons.OK,
                     MessageBoxIcon.Error);
 
@@ -150,6 +142,7 @@ namespace RaspberryDebugger
             }
 
             projectSettings.TargetGroup = targetGroup.Text;
+            projectSettings.UseWebServerProxy = webServerProxyCheck.Checked;
 
             DialogResult = DialogResult.OK;
         }
@@ -158,7 +151,7 @@ namespace RaspberryDebugger
         /// Handles Cancel button clicks.
         /// </summary>
         /// <param name="sender">The sender</param>
-        /// <param name="args"The arguments</param>
+        /// <param name="args">Event arguments</param>
         private void cancelButton_Click(object sender, EventArgs args)
         {
             DialogResult = DialogResult.Cancel;
