@@ -15,11 +15,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-namespace RaspberryDebugger
+namespace RaspberryDebugger.Models.Connection
 {
     /// <summary>
     /// Enumerates the supported SSH authentication types.
@@ -30,12 +28,12 @@ namespace RaspberryDebugger
         /// Password based authentication.
         /// </summary>
         [EnumMember(Value = "password")]
-        Password = 0,
+        PASSWORD = 0,
 
         /// <summary>
         /// Public SSH key based authentication.
         /// </summary>
         [EnumMember(Value = "public-key")]
-        PublicKey
+        PUBLIC_KEY
     }
 }
