@@ -14,18 +14,13 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Configuration;
-using System.Runtime.InteropServices;
-using System.Runtime.Serialization;
-
+using System.Diagnostics.CodeAnalysis;
 using Neon.Net;
 using Newtonsoft.Json;
+using RaspberryDebugger.OptionsPages;
 
-namespace RaspberryDebugger
+namespace RaspberryDebugger.Models.Connection
 {
     /// <summary>
     /// Describes a Raspberry Pi connection's network details and credentials.
@@ -70,11 +65,10 @@ namespace RaspberryDebugger
 
             set
             {
-                if (isDefault != value)
-                {
-                    isDefault = value;
-                    ConnectionsPanel?.ConnectionIsDefaultChanged(this);
-                }
+                if (isDefault == value) return;
+
+                isDefault = value;
+                ConnectionsPanel?.ConnectionIsDefaultChanged(this);
             }
         }
 
@@ -127,7 +121,7 @@ namespace RaspberryDebugger
         /// </summary>
         [JsonProperty(PropertyName = "PrivateKeyPath", Required = Required.Default, DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
         [DefaultValue(null)]
-        public string PrivateKeyPath { get; set; } = null;
+        public string PrivateKeyPath { get; set; }
 
         /// <summary>
         /// Path to the public key for this connection or <c>null</c> when one
@@ -135,7 +129,7 @@ namespace RaspberryDebugger
         /// </summary>
         [JsonProperty(PropertyName = "PublicKeyPath", Required = Required.Default, DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
         [DefaultValue(null)]
-        public string PublicKeyPath { get; set; } = null;
+        public string PublicKeyPath { get; set; }
 
         /// <summary>
         /// Describes the authentication method.
@@ -154,6 +148,7 @@ namespace RaspberryDebugger
         /// appear to have a check box changed event.
         /// </para>
         /// </summary>
+        [SuppressMessage("ReSharper", "InvalidXmlDocComment")]
         internal ConnectionsPanel ConnectionsPanel { get; set; }
     }
 }
