@@ -15,11 +15,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.ComponentModel;
-using System.Net.NetworkInformation;
 
-namespace RaspberryDebugger
+namespace RaspberryDebugger.OptionsPages
 {
     /// <summary>
     /// Implements the custom debug connections options panel.
