@@ -1,3 +1,3 @@
-@using WebApp
+﻿@using WebApp
 @namespace WebApp.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
